Improvments to PackageCacher
